clean 3, file restructure into python package
