--- core/__init__.py
@@ -1,5 +1,0 @@
-"""Core functionality for XAX."""
-
-from xax.core.state import State
-
-__all__ = ["State"]
