--- task/__init__.py
@@ -1,5 +1,0 @@
-"""Task-related functionality for XAX."""
-
-from xax.task.base import BaseTask
-
-__all__ = ["BaseTask"]
