--- utils/__init__.py
@@ -1,5 +1,0 @@
-"""Utility functions for XAX."""
-
-from xax.utils.pytree import PyTree
-
-__all__ = ["PyTree"]
