+++ xax/core/state.py
@@ -1,159 +1,53 @@
-"""Defines a dataclass for keeping track of the current training state."""
+"""State management for XAX."""
 
-import time
-from dataclasses import dataclass
-from typing import Literal, NotRequired, TypedDict, Unpack, cast
+from typing import Any, Dict, Optional
 
-import jax
 import jax.numpy as jnp
-from jaxtyping import Array
-from omegaconf import MISSING
-
-from xax.core.conf import field
-
-Phase = Literal["train", "valid"]
 
 
-def _phase_to_int(phase: Phase) -> int:
-    return {"train": 0, "valid": 1}[phase]
-
-
-def _int_to_phase(i: int) -> Phase:
-    if i < 0 or i > 1:
-        raise ValueError(f"Invalid phase: {i}")
-    return cast(Phase, ["train", "valid"][i])
-
-
-class StateDict(TypedDict, total=False):
-    num_steps: NotRequired[int | Array]
-    num_samples: NotRequired[int | Array]
-    num_valid_steps: NotRequired[int | Array]
-    num_valid_samples: NotRequired[int | Array]
-    start_time_s: NotRequired[float | Array]
-    elapsed_time_s: NotRequired[float | Array]
-    valid_elapsed_time_s: NotRequired[float | Array]
-    phase: NotRequired[Phase]
-    _phase: NotRequired[int | Array]
-
-
-@jax.tree_util.register_dataclass
-@dataclass(frozen=True, kw_only=True)
 class State:
-    _int32_arr: Array = field(MISSING, help="Internal array for storing int64 values")
-    _float32_arr: Array = field(MISSING, help="Internal array for storing floating-point values")
-
-    @property
-    def num_steps(self) -> Array:
-        return self._int32_arr[0]
-
-    @property
-    def num_valid_steps(self) -> Array:
-        return self._int32_arr[1]
-
-    @property
-    def num_samples(self) -> Array:
-        return self._float32_arr[0]
-
-    @property
-    def num_valid_samples(self) -> Array:
-        return self._float32_arr[1]
-
-    @property
-    def start_time_s(self) -> Array:
-        return self._float32_arr[2]
-
-    @property
-    def elapsed_time_s(self) -> Array:
-        return self._float32_arr[3]
-
-    @property
-    def valid_elapsed_time_s(self) -> Array:
-        return self._float32_arr[4]
-
-    @property
-    def phase(self) -> Phase:
-        return _int_to_phase(self._int32_arr[2].item())
-
-    @classmethod
-    def init_state(cls) -> "State":
-        return cls(
-            _int32_arr=jnp.array([0, 0, 0], dtype=jnp.int32),
-            _float32_arr=jnp.array([0.0, 0.0, time.time(), 0.0, 0.0], dtype=jnp.float32),
-        )
-
-    @property
-    def training(self) -> bool:
-        return self.phase == "train"
-
-    def replace(self, **kwargs: Unpack[StateDict]) -> "State":
-        int32_arr = self._int32_arr
-        float32_arr = self._float32_arr
-
-        if "num_steps" in kwargs:
-            int32_arr = int32_arr.at[0].set(kwargs["num_steps"])
-        if "num_valid_steps" in kwargs:
-            int32_arr = int32_arr.at[1].set(kwargs["num_valid_steps"])
-
-        if "phase" in kwargs:
-            int32_arr = int32_arr.at[2].set(_phase_to_int(kwargs["phase"]))
-        if "_phase" in kwargs:
-            int32_arr = int32_arr.at[2].set(kwargs["_phase"])
-
-        if "num_samples" in kwargs:
-            float32_arr = float32_arr.at[0].set(kwargs["num_samples"])
-        if "num_valid_samples" in kwargs:
-            float32_arr = float32_arr.at[1].set(kwargs["num_valid_samples"])
-
-        if "start_time_s" in kwargs:
-            float32_arr = float32_arr.at[2].set(kwargs["start_time_s"])
-        if "elapsed_time_s" in kwargs:
-            float32_arr = float32_arr.at[3].set(kwargs["elapsed_time_s"])
-        if "valid_elapsed_time_s" in kwargs:
-            float32_arr = float32_arr.at[4].set(kwargs["valid_elapsed_time_s"])
-
-        return State(
-            _int32_arr=int32_arr,
-            _float32_arr=float32_arr,
-        )
-
-    def to_dict(self) -> dict[str, int | float | str]:
-        return {
-            "num_steps": int(self.num_steps.item()),
-            "num_valid_steps": int(self.num_valid_steps.item()),
-            "num_samples": int(self.num_samples.item()),
-            "num_valid_samples": int(self.num_valid_samples.item()),
-            "start_time_s": float(self.start_time_s.item()),
-            "elapsed_time_s": float(self.elapsed_time_s.item()),
-            "valid_elapsed_time_s": float(self.valid_elapsed_time_s.item()),
-            "phase": str(self.phase),
-        }
-
-    @classmethod
-    def from_dict(cls, **d: Unpack[StateDict]) -> "State":
-        if "phase" in d:
-            d["_phase"] = _phase_to_int(cast(Phase, d.pop("phase")))
-
-        int32_arr = jnp.array(
-            [
-                d.get("num_steps", 0),
-                d.get("num_valid_steps", 0),
-                d.get("_phase", 0),
-            ],
-            dtype=jnp.int32,
-        )
-
-        float32_arr = jnp.array(
-            [
-                d.get("num_samples", 0),
-                d.get("num_valid_samples", 0),
-                d.get("start_time_s", time.time()),
-                d.get("elapsed_time_s", 0.0),
-                d.get("valid_elapsed_time_s", 0.0),
-            ],
-            dtype=jnp.float32,
-        )
-
-        return cls(
-            _int32_arr=int32_arr,
-            _float32_arr=float32_arr,
-        )
+    """Represents the state of a task during execution."""
+    
+    def __init__(self):
+        """Initialize an empty state."""
+        self._data: Dict[str, Any] = {}
+    
+    def get(self, key: str, default: Optional[Any] = None) -> Any:
+        """Get a value from the state.
+        
+        Args:
+            key: The key to retrieve
+            default: The default value to return if the key is not found
+            
+        Returns:
+            The value associated with the key, or the default if not found
+        """
+        return self._data.get(key, default)
+    
+    def set(self, key: str, value: Any) -> None:
+        """Set a value in the state.
+        
+        Args:
+            key: The key to set
+            value: The value to store
+        """
+        self._data[key] = value
+    
+    def update(self, updates: Dict[str, Any]) -> None:
+        """Update multiple values in the state.
+        
+        Args:
+            updates: A dictionary of key-value pairs to update
+        """
+        self._data.update(updates)
+    
+    def __contains__(self, key: str) -> bool:
+        """Check if a key exists in the state.
+        
+        Args:
+            key: The key to check
+            
+        Returns:
+            True if the key exists, False otherwise
+        """
+        return key in self._data
