+++ xax/task/__init__.py
@@ -1,0 +1,5 @@
+"""Task-related functionality for XAX."""
+
+from xax.task.base import BaseTask
+
+__all__ = ["BaseTask"]
