+++ xax/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility functions for XAX."""
+
+from xax.utils.pytree import PyTree
+
+__all__ = ["PyTree"]
